[REV-119] refactor: 개별설문결과 코드 분리
