+++ src/main/java/com/devcourse/ReviewRanger/eachSurveyResult/repository/EachSurveyResultRepository.java
@@ -1,8 +1,8 @@
-package com.devcourse.ReviewRanger.response.repository;
+package com.devcourse.ReviewRanger.eachSurveyResult.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.devcourse.ReviewRanger.response.domain.EachSurveyResult;
+import com.devcourse.ReviewRanger.eachSurveyResult.domain.EachSurveyResult;
 
 public interface EachSurveyResultRepository extends JpaRepository<EachSurveyResult, Long> {
 }
